pyproject.toml
[project]
name = "holdings-tracker-desktop"
version = "0.1.0"
description = "Desktop application for tracking investment holdings."
authors = [
    {name = "Marco Aurélio Camargo Oliveira", email = "macamargooliveira@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "python-dotenv (>=1.2.1,<2.0.0)",
    "sqlalchemy (>=2.0.44,<3.0.0)"
]

[tool.poetry]
packages = [{include = "holdings_tracker_desktop", from = "src"}]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
